kernel - Implement sbrk(), change low-address mmap hinting (2)
* Fix sysproto.h for the sbrk() change. 'incr' is size_t now.
  (make sysent hadn't been run)

Reported-by: swildner

+++ sys/sys/sysproto.h
@@ -418,13 +418,13 @@ struct	sbrk_args {
 #ifdef _KERNEL
 	struct sysmsg sysmsg;
 #endif
-	int	incr;	char incr_[PAD_(int)];
+	size_t	incr;	char incr_[PAD_(size_t)];
 };
 struct	sstk_args {
 #ifdef _KERNEL
 	struct sysmsg sysmsg;
 #endif
-	int	incr;	char incr_[PAD_(int)];
+	size_t	incr;	char incr_[PAD_(size_t)];
 };
 struct	munmap_args {
 #ifdef _KERNEL
